requirements.txt
streamlit>=1.30.0
google-generativeai>=0.8.0
pypdf>=4.0.0
googlesearch-python
python-whois
python-dotenv
watchdog
